rearrange component folders
