+++ .storybook/assets/avif-test-image.avif
