+++ app/ui/components/header.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/ui/button';
+import { Button } from '@/component-library/button';
 import React from 'react';
 import Link from 'next/link';
 import Image from 'next/image';
