+++ component-library/badge/index.tsx
