+++ component-library/button/Button.stories.tsx
@@ -1,0 +1,49 @@
+import React from 'react';
+import { Meta, StoryFn } from '@storybook/react';
+import { Button, ButtonProps } from '@/component-library/button';
+
+export default {
+  title: 'Components/Button',
+  component: Button,
+  argTypes: {
+    variant: {
+      control: {
+        type: 'select',
+        options: ['primary', 'secondary', 'danger'],
+      },
+    },
+    onClick: { action: 'clicked' },
+  },
+} as Meta;
+
+const Template: StoryFn<ButtonProps> = (args) => <Button {...args} />;
+
+export const Default = Template.bind({});
+Default.args = {
+  variant: 'default',
+  children: 'Default Button',
+};
+
+export const Secondary = Template.bind({});
+Secondary.args = {
+  variant: 'secondary',
+  children: 'Secondary Button',
+};
+
+export const Destructive = Template.bind({});
+Destructive.args = {
+  variant: 'destructive',
+  children: 'Destructive Button',
+};
+
+export const Ghost = Template.bind({});
+Ghost.args = {
+  variant: 'ghost',
+  children: 'Ghost Button',
+};
+
+export const Link = Template.bind({});
+Link.args = {
+  variant: 'link',
+  children: 'A link',
+};
