+++ tailwind.config.ts
@@ -4,7 +4,7 @@ const config: Config = {
   darkMode: ['class', '[data-mode="dark"]'],
   content: [
     './pages/**/*.{js,ts,jsx,tsx,mdx}',
-    './components/**/*.{js,ts,jsx,tsx,mdx}',
+    './component-library/**/*.{js,ts,jsx,tsx,mdx}',
     './app/**/*.{js,ts,jsx,tsx,mdx}',
     './.storybook/**/*.{js,ts,jsx,tsx}', // Include Storybook stories
   ],
